fix(mcp): use i32 for SELECT 1 result in is_client_approved
PostgreSQL returns INT4 (32-bit) for integer literals like SELECT 1,
but the Rust code expected i64 (INT8). This caused a type mismatch
error during OAuth callback when checking client approval status.

Fixes #24

+++ mcp/src/oauth/store.rs
@@ -816,7 +816,7 @@ impl TokenStore {
 
     /// Returns true if the given user has approved the given OAuth client.
     pub async fn is_client_approved(&self, user_id: Uuid, client_id: &str) -> Result<bool> {
-        let exists: Option<(i64,)> = sqlx::query_as(
+        let exists: Option<(i32,)> = sqlx::query_as(
             r#"
             SELECT 1
             FROM mcp_oauth.approved_clients
